requirements.txt
django
djangorestframework
django-rest-swagger
dj-database-url
gunicorn
postgres
psycopg2-binary
pytz
whitenoise
django-heroku
upstox
python-dateutil
django-cors-headers
channels
asgi_redis
